% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutRain.r
\name{cutRain}
\alias{cutRain}
\title{Cut and save wav files according to rain classification}
\usage{
cutRain(x, threshold = c("min", "Q2"), outF, inF, start, label,
  label.only = F)
}
\arguments{
\item{x}{A matrix of results from \code{classifyRain}}

\item{threshold}{A character vector with the threshold to use either "min" or "Q2". Defaults to "min"}

\item{outF}{Destination folder for new wav files. Defaults to home directory if missing.}

\item{inF}{Source folder for wav files. Defaults to home directory if missing.}

\item{start}{Optional. A vector of dateTime objects giving the start time for each file in class.results}

\item{label}{Optional. A character vector, either "audacity" or "raven", to include a label file to
be written in \code{outF} for the original wav, labelling all non-rain sections (ie those that will be exported)}
}
\value{
A dataframe detailing the new wav files created, with, file path, start and end times
logical columns with results of each threshold classification
}
\description{
Using results from the \code{classifyRain} function, input wav files will be cut and saved according to classification results. Contiguous snippets of non-rain will be saved as new wav files
}
\details{
This function will take results from \code{classifyRain} and join together contiguous snippets (in snippets of lengths given in the t.step argument).
}
\examples{

}
