% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardRain.r
\docType{package}
\name{hardRain}
\alias{hardRain}
\title{hardRain: A package for detecting rain events in audio data}
\description{
The hardRain package provides two main functions:
getThreshold() for calculating PSD and signal to noise thresholds on known rain data, and
classifyRain() which uses the thresholds to identify rain events in other audio data files. A further function,
 cutRain() will label rain events in audio software (e.g. Audacity), and optionally, save contiguous non-rain audio
 as new .wav files.
}
