% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutRain.r
\name{cutRain}
\alias{cutRain}
\title{Cut and save wav files according to rain classification}
\usage{
cutRain(x, threshold = c("min", "Q2"), inF, label.only = F, outF,
  label.type = "audacity")
}
\arguments{
\item{x}{A matrix of results from \code{classifyRain}}

\item{threshold}{A character vector with the threshold type to use ("min" or "Q2"). Defaults to "min"}

\item{inF}{Source folder for wav files. Defaults to home directory if missing.}

\item{label.only}{Logical. If TRUE, only label file is written, wav files are not cut and saved.}

\item{outF}{Destination folder for new wav files. Defaults to inF if missing.}

\item{label.type}{Optional. A character vector, for the moment, just "audacity", to include a label file to
be written in \code{inF} for the original wav, labelling all non-rain sections (ie those that will be exported)}
}
\value{
A dataframe (invisibly) detailing the new wav files created, with, filename, full path,
start times (optionally)
}
\description{
A follow on function to \code{classifyRain} for longer wav files. This function labels and
cuts (removes) sections of hard rain in audio files, then saves contiguous periods without rainfall as
new wav files.
}
\details{
This function is only designed to work with results from \code{classifyRain} where input audio files
are subdivided into smaller time segments with \code{t.step} argument. If \code{label.only} is \code{FALSE},
then segments containing rain will be cut from original audio files, and remaining contiguous audio
saved as new wav files. Optionally, a label file for the original wav can be written to the input folder,
marking segments with rain/clear in the wav file (either for Audacity or Raven software).

If a vector of start times and dates (as POSIX* format) is given in \code{start}, then filenames will be
suffixed with new start times. Otherwise, new wav files are suffixed with a numeric ID for each separate
segment originating from the input wav file.
}
\examples{

}
