% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetrics.r
\name{getMetrics}
\alias{getMetrics}
\title{Get PSD and Signal to Noise ratio for audio files}
\usage{
getMetrics(wav, freqLo = c(0.6, 4.4), freqHi = c(1.2, 5.6),
  t.step = NULL, parallel = F)
}
\arguments{
\item{wav}{A vector of wav filenames (including directories)}

\item{freqLo}{A numeric vector of Lower frequency cut offs for each band - defaults
to 2 bands (0.6-1.2 kHz and 4.4-5.6 kHz)}

\item{freqHi}{A numeric vector of Higher frequency cut off - defaults to 2 bands: (0.6-1.2 kHz and 4.4-5.6 kHz)}

\item{t.step}{NULL or a numeric vector giving time in seconds in which to divide
longer files. If NULL, it is assumed that all files analysed are suitably short (e.g. 15 s each)
and do not need to be subdivided (see details)}

\item{parallel}{Logical. Whether to use multicore processing with the parallel package
(must be loaded)}
}
\value{
A numeric matrix with columns \code{psd} and \code{s2n} for each wav file in \code{wav},
filenames are conserved in the rownames
}
\description{
This function does not generally need to be called directly. It is the workhorse function that
reads wav files, extracts psd and signal to noise for specified frequency bands using seewave
functions spec() or meanspec(). This function is called by getThreshold() and classifyRain() which will generally
be used directly.
}
\examples{
See examples in getThreshold() and \\code{\\link{classifyRain}}
}
