% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getThreshold.r
\name{getThreshold}
\alias{getThreshold}
\title{Calculate minimum and 2nd quartile thresholds for rain detection in audio files}
\usage{
getThreshold(wav, freqLo = c(0.6, 4.4), freqHi = c(1.2, 5.6),
  fn = c("meanspec", "spec"), parallel = F)
}
\arguments{
\item{wav}{A vector of wav filenames (including directories)}

\item{freqLo}{a vector of Lower frequency cut offs - defaults to 2 bands (0.6-1.2 kHz and 4.4-5.6 kHz)}

\item{freqHi}{a vector of Higher frequency cut off - defaults to 2 bands: (0.6-1.2 kHz and 4.4-5.6 kHz)}

\item{fn}{a character vector, which seewave function to use: spec or meanspec (see details)}

\item{parallel}{Logical. Whether to use multicore processing with the parallel package (must be loaded)}
}
\value{
a matrix of \code{min} and \code{Q2} thresholds (rows) for \code{psd} and \code{s2n} at each band (columns)
in \code{freqLo} and \code{freqHi}
}
\description{
Calculate minimum and 2nd quartile thresholds for rain detection in audio files
}
\examples{

# Get filenames of training data (known rain recordings in wav files). Only five files are used here for purposes
of this example
train.fn <- list.files(system.file("extdata/rain", package = "hardRain"), "\\\\.wav$", full.names = T)

# Calculate the threshold using default settings - for two frequency bands
trBR <- getThreshold(train.fn, fn = "spec")
trBR
}
