% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardRain.r
\docType{package}
\name{hardRain}
\alias{hardRain}
\alias{hardRain-package}
\title{hardRain: A package for detecting rain events in audio data}
\description{
The hardRain package provides two main functions:
getThreshold() for calculating PSD and signal to noise thresholds on known rain data, and
classifyRain() which uses the thresholds to identify rain events in other audio data
}
